LPD-73438 Delete only if there is CTSContent

+++ modules/apps/change-tracking/change-tracking-store-service/src/main/java/com/liferay/change/tracking/store/internal/CTStore.java
@@ -25,6 +25,7 @@ import java.util.Collection;
 import java.util.Comparator;
 import java.util.HashMap;
 import java.util.HashSet;
+import java.util.List;
 import java.util.Map;
 import java.util.Set;
 
@@ -89,8 +90,13 @@ public class CTStore implements Store {
 			}
 		}
 
-		_ctsContentLocalService.deleteCTSContentsByDirectory(
-			companyId, repositoryId, dirName, _storeType);
+		List<CTSContent> ctContents = _ctsContentLocalService.getCTSContentsByDirectory(
+				companyId, repositoryId, dirName, _storeType);
+
+		if (!ctContents.isEmpty()) {
+			_ctsContentLocalService.deleteCTSContentsByDirectory(
+				companyId, repositoryId, dirName, _storeType);
+		}
 	}
 
 	@Override
